pyproject.toml
[project]
name = "document-intelligence-refinery"
version = "0.1.0"
description = "Document Intelligence Refinery: Agentic pipelines for unstructured document extraction."
authors = [{name = "Ermiyas",  email = "ermiyasbitew239@gmail.com"}]
requires-python = ">=3.8"

[tool.poetry.dependencies]
pdfplumber = "^0.11.0"
pydantic = "^2.6.4"
pytest = "^8.1.1"
pyyaml = "^6.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
